--- frontend/src/pages/apps/user465_projects.svelte
@@ -1,6 +1,0 @@
-<script>
-    import MyProjects from './_components/Projects.svelte';
-
-</script>
-
-<MyProjects></MyProjects>
